Moved almost all canvas things to DrawUtils
