+++ src/AtomMode.ts
@@ -5,9 +5,11 @@
 
 import {Point} from "./AEG/Point";
 import {AtomNode} from "./AEG/AtomNode";
-import {redrawCut, tree} from "./index";
+import {tree} from "./index";
+import {redrawTree} from "./DrawUtils";
 import {offset} from "./DragMode";
 import {legalColor, illegalColor} from "./Themes";
+import {drawAtom} from "./DrawUtils";
 
 //Setting Up Canvas
 const canvas: HTMLCanvasElement = <HTMLCanvasElement>document.getElementById("canvas");
@@ -19,11 +21,6 @@ const ctx: CanvasRenderingContext2D = res;
 
 //HTML letter display
 const atomDisplay = <HTMLParagraphElement>document.getElementById("atomDisplay");
-
-//HTML bounding box check
-const atomCheckBox = <HTMLInputElement>document.getElementById("atomBox");
-const atomCheckBoxes = <HTMLInputElement>document.getElementById("atomBoxes");
-atomCheckBoxes.addEventListener("input", checkBoxRedraw);
 
 //Allows font measurement in pixels to creature atom bounding box.
 let atomMetrics: TextMetrics;
@@ -62,8 +59,7 @@ export function atomMouseDown(event: MouseEvent) {
         atomMetrics.fontBoundingBoxDescent + atomMetrics.actualBoundingBoxAscent
     );
 
-    ctx.clearRect(0, 0, canvas.width, canvas.height);
-    redrawCut(tree.sheet, offset);
+    redrawTree(tree);
     if (tree.canInsert(currentAtom)) {
         drawAtom(currentAtom, legalColor(), true);
     } else {
@@ -83,8 +79,7 @@ export function atomMouseMove(event: MouseEvent) {
         atomMetrics.fontBoundingBoxDescent + atomMetrics.actualBoundingBoxAscent
     );
 
-    ctx.clearRect(0, 0, canvas.width, canvas.height);
-    redrawCut(tree.sheet, offset);
+    redrawTree(tree);
     if (!wasOut) {
         if (tree.canInsert(currentAtom)) {
             drawAtom(currentAtom, legalColor(), true);
@@ -108,8 +103,7 @@ export function atomMouseUp(event: MouseEvent) {
     if (tree.canInsert(currentAtom) && !wasOut) {
         tree.insert(currentAtom);
     }
-    ctx.clearRect(0, 0, canvas.width, canvas.height);
-    redrawCut(tree.sheet, offset);
+    redrawTree(tree);
 }
 
 /**
@@ -117,37 +111,5 @@ export function atomMouseUp(event: MouseEvent) {
  */
 export function atomMouseOut() {
     wasOut = true;
-    ctx.clearRect(0, 0, canvas.width, canvas.height);
-    redrawCut(tree.sheet, offset);
-}
-
-/**
- * Draws the given atomNode with the given color.
- * @param thisAtom the atomMode to be drawn.
- * @param color the color of the atom.
- */
-export function drawAtom(thisAtom: AtomNode, color: string, currentAtom: Boolean) {
-    ctx.textBaseline = "bottom";
-    atomMetrics = ctx.measureText(thisAtom.identifier);
-    ctx.fillStyle = color;
-    ctx.strokeStyle = color;
-    ctx.beginPath();
-    ctx.fillText(thisAtom.identifier, thisAtom.origin.x + offset.x, thisAtom.origin.y + offset.y);
-    if (atomCheckBoxes.checked || (atomCheckBox.checked && currentAtom)) {
-        ctx.rect(
-            thisAtom.origin.x + offset.x,
-            thisAtom.origin.y + offset.y - atomMetrics.actualBoundingBoxAscent,
-            thisAtom.width,
-            thisAtom.height
-        );
-    }
-    ctx.stroke();
-}
-
-/**
- * When the checkbox for showing all bounding boxes is checked redraws the canvas showing the boxes.
- */
-function checkBoxRedraw() {
-    ctx.clearRect(0, 0, canvas.width, canvas.height);
-    redrawCut(tree.sheet, offset);
+    redrawTree(tree);
 }
