+++ src/CopyMultiMode.ts
@@ -7,19 +7,11 @@
 import {Point} from "./AEG/Point";
 import {AtomNode} from "./AEG/AtomNode";
 import {CutNode} from "./AEG/CutNode";
-import {redrawCut, tree} from "./index";
+import {tree} from "./index";
 import {offset} from "./DragMode";
-import {drawAtom} from "./AtomMode";
+import {drawAtom, redrawTree} from "./DrawUtils";
 import {legalColor, illegalColor} from "./Themes";
 import {validateChildren, drawAltered, insertChildren, alterAtom} from "./EditModeUtils";
-
-//Setting Up Canvas
-const canvas: HTMLCanvasElement = <HTMLCanvasElement>document.getElementById("canvas");
-const res: CanvasRenderingContext2D | null = canvas.getContext("2d");
-if (res === null) {
-    throw Error("2d rendering context not supported");
-}
-const ctx: CanvasRenderingContext2D = res;
 
 //The initial point the user pressed down.
 let startingPoint: Point;
@@ -58,8 +50,7 @@ export function copyMultiMouseMove(event: MouseEvent) {
             event.y - startingPoint.y
         );
 
-        ctx.clearRect(0, 0, canvas.width, canvas.height);
-        redrawCut(tree.sheet, offset);
+        redrawTree(tree);
         if (currentNode instanceof CutNode) {
             if (validateChildren(currentNode, moveDifference)) {
                 drawAltered(currentNode, legalColor(), moveDifference);
@@ -104,8 +95,7 @@ export function copyMultiMouseUp(event: MouseEvent) {
             }
         }
     }
-    ctx.clearRect(0, 0, canvas.width, canvas.height);
-    redrawCut(tree.sheet, offset);
+    redrawTree(tree);
     legalNode = false;
 }
 
@@ -115,6 +105,5 @@ export function copyMultiMouseUp(event: MouseEvent) {
  */
 export function copyMultiMouseOut() {
     legalNode = false;
-    ctx.clearRect(0, 0, canvas.width, canvas.height);
-    redrawCut(tree.sheet, offset);
+    redrawTree(tree);
 }
