+++ src/DeleteSingleMode.ts
@@ -8,17 +8,9 @@ import {Point} from "./AEG/Point";
 import {AtomNode} from "./AEG/AtomNode";
 import {CutNode} from "./AEG/CutNode";
 import {offset} from "./DragMode";
-import {drawAtom} from "./AtomMode";
-import {drawCut} from "./CutMode";
-import {redrawCut, tree} from "./index";
+import {drawAtom, drawCut, redrawTree} from "./DrawUtils";
+import {tree} from "./index";
 import {illegalColor} from "./Themes";
-
-const canvas: HTMLCanvasElement = <HTMLCanvasElement>document.getElementById("canvas");
-const res: CanvasRenderingContext2D | null = canvas.getContext("2d");
-if (res === null) {
-    throw Error("2d rendering context not supported.");
-}
-const ctx: CanvasRenderingContext2D = res;
 
 //The initial point the user pressed down.
 let startingPoint: Point;
@@ -61,8 +53,7 @@ export function deleteSingleMouseMove(event: MouseEvent) {
     const newNode: CutNode | AtomNode | null = tree.getLowestNode(newPoint);
     if (currentNode !== null && currentNode !== newNode) {
         legalNode = true;
-        ctx.clearRect(0, 0, canvas.width, canvas.height);
-        redrawCut(tree.sheet, offset);
+        redrawTree(tree);
         if (newNode === tree.sheet || newNode === null) {
             currentNode = null;
             legalNode = false;
@@ -95,8 +86,7 @@ export function deleteSingleMouseUp(event: MouseEvent) {
             }
         }
     }
-    ctx.clearRect(0, 0, canvas.width, canvas.height);
-    redrawCut(tree.sheet, offset);
+    redrawTree(tree);
     currentNode = null;
     legalNode = false;
 }
